+++ core-navigation/conf/routes
@@ -26,4 +26,6 @@ GET    /front-trails/*path.json                             controllers.MoreStor
 GET    /front-trails/*path                                  controllers.MoreStoriesController.renderFrontTrails(path)
 
 # Experimental
-GET    /wikipedia/*path                                     controllers.WikipediaController.render(path)
+GET    /cards/opengraph/*path                               controllers.CardController.opengraph(path)
+GET    /cards/wikipedia/*path                               controllers.CardController.wikipedia(path)
+
